Rename LibEbpfEvents to EbpfEvents
CMake will automatically prepend "Lib" to the name of something created
with add_library, resulting in a artifact called "LibLibEbpfEvents",
this has been bugging me for some time.

Also rename "LibEbpfEvents" directory to simply "Lib" for consistency
with HostIsolation directory, which puts all its userspace components in
a "Lib" directory.
