+++ non-GPL/Events/CMakeLists.txt
@@ -4,6 +4,5 @@
 # or more contributor license agreements. Licensed under the Elastic License 2.0;
 # you may not use this file except in compliance with the Elastic License 2.0.
 
-add_subdirectory(LibEbpfEvents)
+add_subdirectory(Lib)
 add_subdirectory(EventsTrace)
-add_subdirectory(LibEbpfEventsTestLoader)
