+++ non-GPL/Events/EventsTrace/CMakeLists.txt
@@ -5,12 +5,13 @@
 # you may not use this file except in compliance with the Elastic License 2.0.
 
 add_executable(EventsTrace EventsTrace.c)
-add_dependencies(EventsTrace LibEbpfEvents EventProbe_skeleton libbpf libelf)
-target_link_libraries(EventsTrace LibEbpfEvents ${LIBBPF_LIB} ${LIBELF_LIB} -lz)
+add_dependencies(EventsTrace EbpfEvents EventProbe_skeleton libbpf libelf)
+target_link_libraries(EventsTrace EbpfEvents ${LIBBPF_LIB} ${LIBELF_LIB} -lz)
 
 if (BUILD_STATIC_EVENTSTRACE)
     target_link_libraries(EventsTrace -static)
 endif()
+
 target_include_directories(EventsTrace PRIVATE
     "${LIBBPF_TARGET_INCLUDE_DIR}"
     "${LIBBPF_UAPI_INCLUDE_DIR}"
