+++ non-GPL/Events/EventsTrace/EventsTrace.c
@@ -21,7 +21,7 @@
 #include <netinet/in.h>
 
 #define __aligned_u64 __u64 __attribute__((aligned(8)))
-#include <LibEbpfEvents.h>
+#include <EbpfEvents.h>
 
 const char *argp_program_version     = "EventsTrace 0.0.0";
 const char *argp_program_bug_address = "https://github.com/elastic/ebpf/issues";
