+++ non-GPL/Events/Lib/TestLoader/CMakeLists.txt
@@ -4,11 +4,11 @@
 # or more contributor license agreements. Licensed under the Elastic License 2.0;
 # you may not use this file except in compliance with the Elastic License 2.0.
 
-add_executable(LibEbpfEventsTestLoader LibEbpfEventsTestLoader.c)
-add_dependencies(LibEbpfEventsTestLoader LibEbpfEvents EventProbe_skeleton libbpf libelf)
-target_link_libraries(LibEbpfEventsTestLoader LibEbpfEvents ${LIBBPF_LIB} ${LIBELF_LIB} -lz)
+add_executable(EbpfEventsTestLoader EbpfEventsTestLoader.c)
+add_dependencies(EbpfEventsTestLoader EbpfEvents EventProbe_skeleton libbpf libelf)
+target_link_libraries(EbpfEventsTestLoader EbpfEvents ${LIBBPF_LIB} ${LIBELF_LIB} -lz)
 
-target_include_directories(LibEbpfEventsTestLoader PRIVATE
+target_include_directories(EbpfEventsTestLoader PRIVATE
     "${LIBBPF_TARGET_INCLUDE_DIR}"
     "${LIBBPF_UAPI_INCLUDE_DIR}"
     "${LIBBPF_INCLUDE_DIR}"
